+++ workers/ohsome_quality_analyst/api/request_models.py
@@ -38,7 +38,8 @@ class BaseIndicator(BaseModel):
     )
     include_svg: bool = False
     include_html: bool = False
-    flatten: bool = True
+    include_data: bool = False
+    flatten: bool = False
 
     class Config:
         """Pydantic config class."""
@@ -49,10 +50,13 @@ class BaseIndicator(BaseModel):
 
 
 class BaseReport(BaseModel):
-    name: ReportEnum = pydantic.Field(..., title="Report Name", example="SimpleReport")
+    name: ReportEnum = pydantic.Field(
+        ..., title="Report Name", example="BuildingReport"
+    )
     include_svg: bool = False
     include_html: bool = False
-    flatten: bool = True
+    include_data: bool = False
+    flatten: bool = False
 
     class Config:
         """Pydantic config class."""
@@ -188,7 +192,7 @@ INDICATOR_EXAMPLES = {
             "fidField": "ogc_fid",
             "includeSvg": False,
             "includeHtml": False,
-            "flatten": True,
+            "flatten": False,
         },
     },
     "Custom AOI": {
@@ -234,9 +238,6 @@ INDICATOR_EXAMPLES = {
                         ]
                     ],
                 },
-                "includeSvg": False,
-                "includeHtml": False,
-                "flatten": True,
             },
             "layer": {
                 "name": "My layer name",
@@ -285,7 +286,7 @@ INDICATOR_EXAMPLES = {
             },
             "includeSvg": False,
             "includeHtml": False,
-            "flatten": True,
+            "flatten": False,
         },
     },
 }
@@ -302,7 +303,7 @@ REPORT_EXAMPLES = {
             "fidField": "ogc_fid",
             "includeSvg": False,
             "includeHtml": False,
-            "flatten": True,
+            "flatten": False,
         },
     },
     "Custom AOI": {
@@ -326,7 +327,7 @@ REPORT_EXAMPLES = {
             },
             "includeSvg": False,
             "includeHtml": False,
-            "flatten": True,
+            "flatten": False,
         },
     },
 }