+++ workers/ohsome_quality_analyst/utils/helper.py
@@ -1,6 +1,5 @@
 """Standalone helper functions."""
 
-import asyncio
 import importlib
 import json
 import logging
@@ -8,12 +7,10 @@ import os
 import pathlib
 import pkgutil
 import re
-import warnings
 from datetime import date, datetime
-from typing import Coroutine, Generator, Union
+from typing import Generator, Union
 
 import geojson
-import joblib
 import numpy as np
 from geojson import Feature, FeatureCollection, MultiPolygon, Polygon
 
@@ -168,23 +165,3 @@ def flatten_sequence(input_seq: Union[dict, list, tuple, set]) -> list:
         else:
             output.append(val)
     return output
-
-
-def load_sklearn_model(path: str):
-    """Load sklearn model from disk
-
-    Raise an error if a `UserWarning` is thrown during loading of a model from disk.
-    The `UserWarning` is most likely due to use of different versions of `scikit-learn`
-    for dumping and for loading the model with `joblib`.
-    https://scikit-learn.org/stable/modules/model_persistence.html#security-maintainability-limitations
-    """
-    warnings.simplefilter("error", UserWarning)  # Raise exception if warning occurs
-    model = joblib.load(path)
-    warnings.resetwarnings()
-    return model
-
-
-async def sem_task(task: Coroutine, semaphore=asyncio.Semaphore(4)) -> Coroutine:
-    """Run task with semaphore. Semaphore limits num of concurrent executions."""
-    async with semaphore:
-        return await task